pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.64.0",
    "groq>=0.31.0",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "zeep>=4.3.1",
]
